feat: Add README.md
This commit introduces the initial `README.md` file for the project.

The README provides an overview of the "KITE - Remote Compose Demo" project, explaining its purpose as a demonstration of server-driven UI using Remote Compose. It includes a link to an explanatory article, outlines use cases, and gives instructions on how to run the project. It also lists the key Remote Compose dependencies being used.